requirements.txt
numpy
torch
torchvision
Pillow
tqdm
tensorboardX
scikit-learn